% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function to cleans NOAA earthquake and localisation data


Function eq_clean_data()}
\usage{
eq_clean_data(earthquake_track)
}
\arguments{
\item{earthquake_track}{raw dataframe obtained after reading NOAA Significant Earthquake
Database tab-separated file}
}
\value{
This function returns a data frame with clean data ready to make a plot
}
\description{
This function returns a date frame having the following:
1.A date column created by uniting the year, month, day and converting it
to the Date class
2. LATITUDE and LONGITUDE columns converted to numeric class
3. converting data to numeric class
}
\examples{
\dontrun{
library(dplyr)
library(magrittr)
library(readr)

earthquake_tracka_clean<-eq_clean_data(earthquake_track)
}


}
