% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function eq_create_label()}
\usage{
eq_create_label(clean_data)
}
\arguments{
\item{clean_data, }{a dataframe obtained after runing functions eq_clean_data()
and eq_location_clean(). Initial raw data should be downloaded from NOAA website}
}
\value{
This function creates HTML label with annotations
}
\description{
This function Create a function takes the dataset as an argument
and creates an HTML label that can be used as the annotation text in the leaflet map.
This function puts together a character string for each earthquake that shows
the cleaned location, the magnitude (EQ_PRIMARY), and the total number of deaths
(TOTAL_DEATHS), with #boldface labels for each ("Location", "Total deaths", and
"Magnitude"). If an earthquake is missing values for any of these, both the label
and the value should be skipped for that element of the tag.
}
\examples{
\dontrun{
clean_data \%>\%
dplyr::filter(COUNTRY == "IRAN" & lubridate::year(DATE) >= 2016) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}


}
