% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function eq_location_clean()}
\usage{
eq_location_clean(earthquake_track_clean)
}
\arguments{
\item{earthquake_track_clean}{obtained after runnign eq_clean_data() function on
NOAA Significant Earthquake Database tab-separated file}
}
\value{
This function returns a data frame with clean location names
}
\description{
This function cleans the LOCATION_NAME column by stripping out
the country name (including the colon) and converts names to title case
(as opposed to all caps)
}
\examples{
\dontrun{
library(dplyr)
library(stringr)
library(magrittr)

earthquake_track_clean<-eq_location_clean(earthquake_track_clean)
}


}
