% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_map.R
\name{eq_map}
\alias{eq_map}
\title{Function to map NOAA earthquake data

Function eq_map()}
\usage{
eq_map(earthquake_track_clean, annot_col)
}
\arguments{
\item{earthquake_track_clean, }{a dataframe obtained after runing functions eq_clean_data()
and eq_location_clean(). Initial raw data should be downloaded from NOAA website}

\item{annot_col}{name of column in the dataframe, that should be used
for annotation}
}
\value{
This function returns a map with earthquake epicentres and annotations
within pop-up window
}
\description{
This function takes an argument data containing
the filtered data frame with earthquakes to visualize. The function maps
the epicentres (LATITUDE/LONGITUDE) and annotates each point with in pop
up window containing annotation data stored in a column of the data frame.
The user can choose which column is used for the annotation
in the pop-up with a function argument named annot_col. Each earthquake
is shown with a circle, and the radius of the circle should be
proportional to the earthquake's magnitude
}
\examples{
\dontrun{
earthquake_track_clean \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE")
}


}
