% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom to plot a time line with points for each earthquake

Geom function geom_timeline()}
\usage{
geom_timeline(mapping = NULL, data = NULL, position = "identity",
  stat = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created}

\item{data}{The data to be displayed}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Other arguments passed on to the layer. it can be used to set an
aesthetic to a fixed value.}
}
\value{
function produce a time line plot of selected earthquakes
}
\description{
a geom for ggplot2 called geom_timeline() for plotting a time line of
earthquakes ranging from xmin to xmax dates with a point for each earthquake.
Aesthetics include colour, size, and alpha (for transparency).
}
\details{
dataframe obtained after running functions eq_clean_data()
and eq_location_clean(). Initial raw data should be downloaded from NOAA website.
The data must be cleaned using the functions eq_clean_data and eq_location_clean,
that are included in the package. Aesthetics can be specified in geom_timeline or
in ggplot
}
\examples{
\dontrun{
earthquake_track_clean \%>\%
dplyr::filter(COUNTRY \%in\% c('JAPAN', 'IRAN',YEAR > 2010)) \%>\%

   #option to filter date:
#dplyr::filter(DATE > '2010-01-01', DATE < '2015-01-01') \%>\%
ggplot() +
geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS, size = EQ_PRIMARY)) +
scale_size_continuous(name = 'Richter scale value') +
scale_color_continuous(name = '# of Deaths')+
theme(legend.position="bottom", legend.box="horizontal")
}


}
