% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timelin_label.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{Geom function GeomTimelinelabel that contains code}
\format{An object of class \code{GeomTimelinelabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelinelabel
}
\description{
Geom function GeomTimelinelabel that contains code
}
\keyword{datasets}
